# Improved constraint deletion performance by separating functions