--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -41,11 +41,8 @@ end
 
     @testset "Linear tests" begin
         config = MOIT.TestConfig(solve=false)
-        MOIT.contlineartest(solver, config, [
-            # partial_start requires VariablePrimalStart to be implemented by the
-            # solver.
-            "partial_start"
-        ])
+        # Exclude partial_start since VariablePrimalStart not implemented.
+        MOIT.contlineartest(solver, config, ["partial_start"])
         config = MOIT.TestConfig()
         include("contlinear.jl")
         set_linear1test_solutions!(solver)