This change improves the performance of constraint deletion by more than 6x.

Using the following benchmark:
```julia
using JuMP
using GLPK
using Clp
using BenchmarkTools

function constraint_deletion_benchmark(opt_package,num_variables,solver_opts=NamedTuple())
    model = Model(with_optimizer(opt_package.Optimizer; solver_opts...))

    vars=@variable(model,[i = 1:num_variables])

    cons=@constraint(model,vars .<= 1)

    for con in cons
        delete(model,con)
    end

    cons=@constraint(model,vars .<= 1)
    @objective(model,Max,sum(vars))
    optimize!(model)

    for con in cons
        delete(model,con)
    end
end

#------
@benchmark constraint_deletion_benchmark(GLPK,1000)
@benchmark constraint_deletion_benchmark(Clp,1000,(LogLevel=0,))

```
master:
```julia
julia> @benchmark constraint_deletion_benchmark(GLPK,1000)
BenchmarkTools.Trial: 
  memory estimate:  78.52 MiB
  allocs estimate:  3852740
  --------------
  minimum time:     295.216 ms (4.87% GC)
  median time:      327.834 ms (4.88% GC)
  mean time:        328.450 ms (6.63% GC)
  maximum time:     416.230 ms (28.00% GC)
  --------------
  samples:          16
  evals/sample:     1

julia> @benchmark constraint_deletion_benchmark(Clp,1000,(LogLevel=0,))
BenchmarkTools.Trial: 
  memory estimate:  78.94 MiB
  allocs estimate:  3854763
  --------------
  minimum time:     289.888 ms (5.43% GC)
  median time:      323.338 ms (4.64% GC)
  mean time:        326.543 ms (6.63% GC)
  maximum time:     423.405 ms (28.99% GC)
  --------------
  samples:          16
  evals/sample:     1
```

ndinsmore:row_deletion_perf:
```julia
julia> @benchmark constraint_deletion_benchmark(GLPK,1000)
BenchmarkTools.Trial: 
  memory estimate:  5.98 MiB
  allocs estimate:  118130
  --------------
  minimum time:     40.884 ms (0.00% GC)
  median time:      50.572 ms (0.00% GC)
  mean time:        60.356 ms (2.37% GC)
  maximum time:     231.075 ms (3.69% GC)
  --------------
  samples:          83
  evals/sample:     1

julia> @benchmark constraint_deletion_benchmark(Clp,1000,(LogLevel=0,))
BenchmarkTools.Trial: 
  memory estimate:  6.41 MiB
  allocs estimate:  120153
  --------------
  minimum time:     30.155 ms (0.00% GC)
  median time:      43.865 ms (0.00% GC)
  mean time:        57.698 ms (3.26% GC)
  maximum time:     515.472 ms (1.47% GC)
  --------------
  samples:          87
  evals/sample:     1
```