@ndinsmore 2019-02-26
```julia
for (key,val) in scalar
        scalar[key] = val > row ? val - 1 : val
end
```
yields
```julia
julia> @benchmark constraint_deletion_benchmark(GLPK,1000)
BenchmarkTools.Trial: 
  memory estimate:  13.62 MiB
  allocs estimate:  618630
  --------------
  minimum time:     91.325 ms (0.00% GC)
  median time:      100.139 ms (2.98% GC)
  mean time:        101.903 ms (2.03% GC)
  maximum time:     128.013 ms (2.51% GC)
  --------------
  samples:          50
  evals/sample:     1
```
-------------------
```julia
for (key, value) in scalar
    if value > row
        scalar[key] -= 1
    end
end
```
yields
```julia
julia> @benchmark constraint_deletion_benchmark(GLPK,1000)
BenchmarkTools.Trial: 
  memory estimate:  21.23 MiB
  allocs estimate:  1117130
  --------------
  minimum time:     139.537 ms (2.12% GC)
  median time:      148.302 ms (1.99% GC)
  mean time:        151.534 ms (1.93% GC)
  maximum time:     173.682 ms (2.35% GC)
  --------------
  samples:          33
  evals/sample:     1
```