@joaquimg 2019-02-27
Fast deleting is great 6x looks a important improvement.
Looks like a bottleneck.
If someone has a better data structure to propose if would be great.
Otherwise accessing this internal field is bad but not that bad...
Are there built-in Dict functions to it?