@odow 2019-02-27
What about
```julia
for (key, value) in scalar
    if value > row
        scalar[key] = value - 1
    end
end
```
This avoids the extra lookup associated with the `-=`. 

We've gone from 300ms to 100ms (Great!) (and probably even a little less if we have to perform less `setindex!` above). I don't think accessing the `.vals` field of a dict offsets the <40ms gain we get.

If this is a performance issue, I would regard this as a problem in Base Julia. It should provide an exported, documented function that allows querying and setting a key-value pair without having to perform multiple lookups.