@ndinsmore 2019-02-27
I have asked this basic question on discourse.  But some investigation shows that the naive solution really starts to fall off above 1000 elements.
![image](https://user-images.githubusercontent.com/45537276/53521799-00454580-3aa7-11e9-8003-55543514d221.png)