@ndinsmore 2019-03-05
This is the discourse post:
https://discourse.julialang.org/t/fast-dict-value-modification-by-accessing-dict-vals/21277/10

Which has evolved into the Julia PR: https://github.com/JuliaLang/julia/pull/31223