@ndinsmore 2019-03-06
It looks like the PR is gaining enough momentum that it will eventually be added.  In the mean time should we implement the naive version?