@odow 2019-03-06
Great, thanks for escalating this to Base Julia. Let's implement the naive loop version with a TODO linking to the Julia PR.