@ndinsmore 2019-03-14
@odow the PR has been merged with master: https://github.com/JuliaLang/julia/pull/31223
I have update this repo to include the definition of map! then to also make use of it. 
The code for map! is the same as that in julia master