@joaquimg 2019-03-14
Slightly related comment:
There is a related open issue https://github.com/JuliaLang/julia/issues/31199
That could also benefit JuMP if you are interested.