@ndinsmore 2019-03-20
Master:
```julia
julia> @benchmark constraint_deletion_benchmark(GLPK,1000)
BenchmarkTools.Trial: 
  memory estimate:  78.49 MiB
  allocs estimate:  3863538
  --------------
  minimum time:     301.017 ms (4.38% GC)
  median time:      337.761 ms (4.04% GC)
  mean time:        350.259 ms (5.62% GC)
  maximum time:     444.934 ms (24.12% GC)
  --------------
  samples:          15
  evals/sample:     1

julia> @benchmark constraint_deletion_benchmark(Clp,1000,(LogLevel=0,))
BenchmarkTools.Trial: 
  memory estimate:  78.72 MiB
  allocs estimate:  3863560
  --------------
  minimum time:     292.421 ms (3.75% GC)
  median time:      320.846 ms (4.76% GC)
  mean time:        377.702 ms (8.67% GC)
  maximum time:     871.552 ms (29.00% GC)
  --------------
  samples:          14
  evals/sample:     1
``
row_deletion_perf:
```julia
julia> @benchmark constraint_deletion_benchmark(GLPK,1000)
BenchmarkTools.Trial: 
  memory estimate:  5.74 MiB
  allocs estimate:  115228
  --------------
  minimum time:     37.724 ms (0.00% GC)
  median time:      42.170 ms (0.00% GC)
  mean time:        46.151 ms (4.75% GC)
  maximum time:     183.727 ms (0.00% GC)
  --------------
  samples:          109
  evals/sample:     1

julia> @benchmark constraint_deletion_benchmark(Clp,1000,(LogLevel=0,))
BenchmarkTools.Trial: 
  memory estimate:  5.97 MiB
  allocs estimate:  115250
  --------------
  minimum time:     24.873 ms (0.00% GC)
  median time:      29.140 ms (0.00% GC)
  mean time:        30.183 ms (4.00% GC)
  maximum time:     51.206 ms (10.70% GC)
  --------------
  samples:          166
  evals/sample:     1
```