@odow 2019-04-11
Merged despite the failing tests on 0.6 because I just merged a PR bumping Julia to >=1.0.